abstract admin application into view files
