+++ .eslintrc
@@ -2,13 +2,19 @@
   "parser": "babel-eslint",
   "extends": "eslint:recommended",
   "env": {
-      "node": true,
-      "mocha": true,
-      "browser": true,
-      "es6": true
+    "node": true,
+    "mocha": true,
+    "browser": true,
+    "es6": true
   },
   "rules": {
-      "strict": 0,
-      "semi": ["error", "always"]
-    }
+    "strict": 0,
+    "semi": [
+      "error",
+      "always"
+    ]
+  },
+  "plugins": [
+    "html"
+  ]
 }
