+++ operations/views/admin/index.html
@@ -3,7 +3,6 @@
   <head>
       <title>node-distribute</title>
       <meta name="viewport" content="width=device-width, initial-scale=1">
-      <link rel="stylesheet" href="/assets/admin.css">
   </head>
   <body>
       <div class="navbar navbar-center">
