+++ operations/views/admin/views/404.vue
@@ -1,0 +1,13 @@
+<template>
+  <pre style="width:425px;height:300px;" class="vertical-center background-black text-white">
+      <h3> This page does not exist </h3>
+      <h1> 404 </h1>
+      <a href="/" class="text-white" style="display:block;"> Go back to somewhere safe </a>
+  </pre>
+</template>
+
+<script>
+module.exports = {
+  data: {}
+};
+</script>
